#2 package structure refactoring
- 레이어 패키지 단위 구조를 도메인 단위의 패키지 구조로 리팩토링 진행
