--- src/main/java/com/voa/goodbam/controller/Controller.java
@@ -1,14 +1,0 @@
-package com.voa.goodbam.controller;
-
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-public class Controller {
-
-    @GetMapping("/")
-    public String index() {
-
-        return "home";
-    }
-
-}
