+++ src/main/java/com/voa/goodbam/notification/NotificationType.java
@@ -1,4 +1,4 @@
-package com.voa.goodbam.domain.notification;
+package com.voa.goodbam.notification;
 
 public enum NotificationType {
     CONFIRM_HOME, CONFIRM_
