+++ src/main/java/com/voa/goodbam/status/domain/UserStatusInRoomRepository.java
@@ -1,11 +1,10 @@
-package com.voa.goodbam.repository;
+package com.voa.goodbam.status.domain;
 
-import com.voa.goodbam.domain.room.UserStatusInRoom;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserStatusInRoomRespository extends CrudRepository<UserStatusInRoom, Long> {
+public interface UserStatusInRoomRepository extends CrudRepository<UserStatusInRoom, Long> {
 
     /**
      *
