+++ src/main/java/com/voa/goodbam/status/type/HomeComingStatus.java
@@ -1,4 +1,4 @@
-package com.voa.goodbam.domain.room;
+package com.voa.goodbam.status.type;
 
 public enum HomeComingStatus {
     NOT_INITIATED, ON_THE_WAY_HOME, ARRIVED_HOME, NEED_CONFIRMATION
